+++ src/Cwaff-Guns/Alyx.cs
@@ -5,7 +5,7 @@ public class Alyx : CwaffGun
     public static string ItemName         = "Alyx";
     public static string ShortDescription = "Welcome to the New Age";
     public static string LongDescription  = "Fires shots that poison and ignite enemies. Current and max ammo decay exponentially, leaving radioactive waste behind in the process. Gun decays completely at 10 max ammo.";
-    public static string Lore             = "A little known fact of nuclear chemistry is that sufficiently large quantities of Uranium -- under specific circumstances not fully understood at present -- can decay directly into guns. These guns have extremely limited lifespans before decaying completely into radiactive goo, but their sheer utility in battle make them a prized treasure for experienced gungeoneers who are willing to absorb a few gamma rays in the name of DPS";
+    public static string Lore             = "A little known fact of nuclear chemistry is that sufficiently large quantities of Uranium -- under specific circumstances not fully understood at present -- can decay directly into guns. These guns have extremely limited lifespans before decaying completely into radioactive goo, but their sheer utility in battle make them a prized treasure for experienced gungeoneers who are willing to absorb a few gamma rays in the name of DPS";
 
     internal const float _AMMO_HALF_LIFE_SECS = 90.0f;
     internal const float _GUN_HALF_LIFE_SECS  = 300.0f;