+++ src/Cwaff-Guns/HolyWaterGun.cs
@@ -5,7 +5,7 @@ public class HolyWaterGun : CwaffGun
     public static string ItemName         = "Holy Water Gun";
     public static string ShortDescription = "Water, Gun, & Holy Soak";
     public static string LongDescription  = "Deals quadruple damage to the Jammed. Killing a Jammed enemy reduces curse by 1.";
-    public static string Lore             = "Rumored to have been used in exorcisms by the High Priest back while he was still the Low Priest. While the exact composition of the holy water is unknown, scientists have been able to reasonably ascertain the fluid contains koi pond water, primer, rat saliva, and moonshine. In any case, it has proven extremely effective at exorcizing the Jammed and nauseating everyone else.";
+    public static string Lore             = "Rumored to have been used in exorcisms by the High Priest back while he was still the Low Priest. While the exact composition of the holy water is unknown, scientists have been able to reasonably ascertain the fluid contains koi pond water, primer, rat saliva, and moonshine. In any case, it has proven extremely effective at exorcising the Jammed and nauseating everyone else.";
 
     internal const float _JAMMED_DAMAGE_MULT = 4f;